+++ test/io.jl
@@ -33,11 +33,12 @@ config = Wflow.Config(tomlpath)
     # modifiers can also be applied
     kvconf = Wflow.get_alias(config.input.vertical, "kv_0", "kv₀", nothing)
     @test kvconf isa Wflow.Config
-    ncname, modifier = Wflow.ncvar_name_modifier(kvconf)
+    ncname, modifier = Wflow.ncvar_name_modifier(kvconf, config = config)
     @test ncname === "KsatVer"
     @test modifier.scale == 1.0
     @test modifier.offset == 0.0
     @test modifier.value === nothing
+    @test modifier.index === nothing
 
     # test the optional "dir_input" and "dir_output" keys
     @test haskey(config, "dir_input")
@@ -252,6 +253,8 @@ end
     @test vertical.cfmax[1] ≈ 3.7565300464630127
     @test vertical.soilthickness[1] ≈ 2000.0
     @test vertical.precipitation[49951] ≈ 2.2100000381469727
+    @test vertical.c[1] ≈
+          [9.152995289601465, 8.919674421902961, 8.70537452585209, 8.690681062890977]
 end
 
 config.input.vertical.cfmax = Dict("value" => 2.0)
@@ -262,6 +265,12 @@ config.input.vertical.soilthickness = Dict(
 )
 config.input.vertical.precipitation =
     Dict("scale" => 1.5, "netcdf" => Dict("variable" => Dict("name" => "precip")))
+config.input.vertical.c = Dict(
+    "scale" => [2.0, 3.0],
+    "offset" => [0.0, 0.0],
+    "layer" => [1, 3],
+    "netcdf" => Dict("variable" => Dict("name" => "c")),
+)
 
 model = Wflow.initialize_sbm_model(config)
 Wflow.load_dynamic_input!(model)
@@ -271,6 +280,12 @@ Wflow.load_dynamic_input!(model)
     @test vertical.cfmax[1] == 2.0
     @test vertical.soilthickness[1] ≈ 2000.0 * 3.0 + 100.0
     @test vertical.precipitation[49951] ≈ 1.5 * 2.2100000381469727
+    @test vertical.c[1] ≈ [
+        2.0 * 9.152995289601465,
+        8.919674421902961,
+        3.0 * 8.70537452585209,
+        8.690681062890977,
+    ]
 end
 
 Wflow.close_files(model, delete_output = false)